refactor(frontend): split Parser into parser/ translation units
Parser.cpp was a single 2280-line file: one class whose methods ranged from
the low-level lexer drivers (advance/eat/peek/synchronize) through the
declaration parsers (class/impl/trait/function/method/constructor/params/
type-params/import/component/using/macro/type-expr) and the statement
parsers (if/while/for/for-in/block/control-body/return/variable-decl/
break/continue/bind-receiver) down to the fifteen-level expression
precedence chain.

The methods stay members of Parser (the call graph and every `this->` access
are untouched) and are now out-of-line definitions in frontend/parser/:

- ParserDecl.cpp   parseClass/parseImpl/parseTrait/parseFunctionDefinition/
                    parseMethod/parseConstructor/parseParams/parseTypeParams/
                    parseImport/parseComponent/parseUsing/parseMacro/
                    parseFunction/parseLambda/parseTypeExpr        (~938L)
- ParserStmt.cpp   parseIfStatement/parseWhileStatement/parseForStatement/
                    parseForInStatement/parseForClause/parseBlock/
                    parseControlBody/parseStatement/parseReturn/
                    parseVariableDeclaration/parseBreakStatement/
                    parseContinueStatement/bindDeclarativeReceiver   (~575L)
- ParserExpr.cpp   parseBaseExpression/parseBoolExpression/
                    parseCoalesceExpression/parseOrExpression/
                    parseAndExpression/parseComparison/parseAdditiveExpression/
                    parseMultiplicativeExpression/parsePowerExpression/
                    parseUnaryExpression/parsePostfix/parsePrimary/
                    parseValue/parseArgs/parseTranspile             (~611L)
- ParserTokens.cpp getTokenName (member) plus the two file-scope helpers
                    getCompoundAssignOp and isAssignableExpr, declared in
                    ParserTokens.h so every parser TU is self-contained
                    (the project's .clangd enforces UnusedIncludes: Strict)
                                                        (~101L)

Parser.cpp drops to ~120 lines (ctor + parse + the lexer drivers).

Like the SemanticAnalyzer split, the member fields stay on the class and the
methods are physically relocated rather than rewritten; this already meets
the plan's file-size and layering goal with zero behavioural risk. The
plan's follow-up "secondary splits" (parseClass -> parseClassMember,
parsePostfix -> parsePostfixSuffix, parsePrimary -> parsePrimaryLiteral/Ident/
Group) are left for a later pass — they require re-cutting function bodies and
were explicitly marked optional, whereas the physical move is what removes the
mega-file.

Behaviour unchanged: 429 tests pass, the 5 failures are the pre-existing
Bedrock-only StdlibObservableTest and the PR#35 ForInTest regression, and the
bytecode dump is byte-identical to the baseline.
